Fix the SQUID_CC_REQUIRE_ARGUMENT autoconf function.
The function tests for the succesful compilation of the test program
with 'if test "x$1" != "xno"', but autoconf will replace $1 with the
first argument of the function.  E.g, if $1 is ac_cv_foo, the result
will be 'if test "xfoo" != "xno"', which is always false.

Correct this by adding another $ in front of the $1 argument.