branch: Keep association with the current context until it is destroyed
Contexts can be tied to complex objects like ALE. Exiting such a context
for the last time triggers a cleanup tsunami. Complex cleanup sequences
are known sources of difficult-to-triage crashes, so it is important to
preserve their context association. Naturally, it is not possible to
preserve the being-destroyed CodeContext object itself, so we extract
and keep the gist of it, hoping that would enough to get the relevant
details from the earlier cache.log debugging or the access.log record.

Refactored Entering() to make sure the destruction happens _before_ we
enter the new context (and to address the above concerns).

Refactored briefCodeContext to enable context gist saving. The new API
will also be used if we add support for sharing the context association
with an SMP worker.