Support TcpAcceptor and ConnStateData/etc code contexts
TcpAcceptor context is just a PortCfg object. We could easily use raw
PortCfg object while processing port configurations, but I have undone
those changes because associated debugs(0/1) messages already detail
their context and polishing some of them is likely to be controversial.
I could not find a good way to compute a PortCfg gist. Perhaps we
should add PortCfg::id. In fact, perhaps all RefCountable objects
should have an InstanceId?!

ConnStateData/etc contexts is its from-client Connection _until_
ConnStateData starts working on a specific transaction (which becomes
the context). When the transaction is over, the context is switched back
to the underlying connection. This support is incomplete because we need
to merge ALE with MasterXaction to polish this further (out of scope).