Streamlined cleanup to reduce accidental self-destruction chances
I suspect doneWithTunnel() design was driven by the desire to reduce
code duplication -- "if done, destroy" logic is otherwise repeated in
connection closure handlers because TunnelStateData is not a job.
However, any "delete this" method is inherently very dangerous. It is
best to make it as simple and as explicit as possible. Here, that meant
renaming it and making the destruction (inside the method) unconditional
(at the expense of that minor logic duplication).

Also renamed doneWithX() methods. There are some unfortunate exceptions,
but most of Squid code is using doneWithFoo() and similar names for
read-only "Am I done with foo?" checks rather than foo's end handlers. I
think that (former) usage is better.