branch: Do not mishandle a protocol called "other" (if any)
Squid configuration uses OTHER to mark rules for all unnamed protocols.
If an admin has to deal with a protocol actually caller "other", they
can still configure other-specific rules because OTHER must be spelled
with all-caps while real protocol comparison is case-insensitive.

However, when Squid receives an Upgrade header that has a item spelled
"other", and there is a rule explicitly naming that protocol, then Squid
must not accidentally apply the rules for unnamed protocols to that
item, and must apply the rule explicitly naming that protocol.

This change also optimizes (possibly common) handling of OTHER protocols
because Squid no longer has to search for "OTHER" in the rules map.