branch: Fixed http_upgrade_request_protocols interpretation
We used to group/merge rules by their protocol names, which violated the
documentation and intent because (a, b, a) rules would be checked
in the (a, a, b) order. Also, the grouping was case sensitive.

There is a (very complex) way we could merge same-protocol rules and
still honor evaluation order and version differences, but the
explicit-iteration implementation being committed is a lot simpler and
probably as efficient.

Removing rule grouping/merging required completely rewriting rules
storage and iteration code.