Moved "close on idle" functionality from custom Dialer to Connection
There are (or will be) other asynchronous calls that need to pass an
orphan (i.e. not owned by any code and not registered with Comm)
connection around. Moreover, I suspect that, eventually, all Connections
will close automatically without referring to Squid bug 3329 because
manual closure is error-prone. That future may be far away, but a
reusable API can help us gauge the amount of code ready for that future.

I have also tried using std::unique_ptr to close the idle Connection
automatically as suggested in the now-addressed XXX, but std::unique_ptr
is not compatible with reference counting. Connection already has (and
uses!) the "close on idle" mechanism. The only missing bit was the
level-4 Squid bug 3329 warning that is not applicable to the new async
call use case. The new COMM_ORPHANED flag marks such use cases.