Ignore unattached Transients entries
After a reply has been cached on shared_memory and(or) disk and the
transaction finishes, the corresponding Transients entry is not freed.
Subsequent requests (with different public keys) may occupy the same
shared memory and rock slots (overriding the already cached entry),
making the initial Transients entry 'unattached'. Requests to
the initially cached resource now stall, because having received the
available Transients entry the Squid worker keeps waiting for a
notification from another worker, incorrectly assuming that this worker
(i.e., the entry writer) exists and is going to fetch it. Since in
reality there is no such writer, the transaction hangs and is aborted
later after timeout.

In order to avoid this effect, we should not return such 'unattached'
Transients entries. This means that an existing Transients entry may
be obtained and returned to the caller if and only if either this entry
has been already anchored to a cache or it has a writer (i.e., another
worker is about to create that cache soon).