branch: Do not leave an idle tunnel hanging after the client is gone
When client connection closure handler is called, and the tunnel has
nothing to do, we must destroy the tunnel. One way to do that is to
close the server connection that will trigger the server closing
handler which, in turn, should deleteThis().

This change also restores the essential handleServerClosure(),
handleClientClosure() symmetry lost in branch commit c94fd17.

XXX: How do we know that server.conn is not nil?!

TODO: It would be better to avoid the round trip to the Comm closing
code -- we know we want to close the server connection and destroy the
job; we do not need to abuse Comm to remind us about it later.