branch: Resolving connection closure methods mess, part 1
handleClientClosure() and handleServerClosure() were named using the
same naming scheme but have orthogonal purposes (besides the clear
client-vs-server difference). The former is a connection closure
handler; a reactor to an external event. The latter is meant to initiate
a tunnel destruction from within.

serverClosed() was named to mimic FwdState::serverClosed() method, which
is important for the upcoming merger of the two classes. Thus, I renamed
handleClientClosure() to clientClosed() to keep the naming symmetry
without changing the serverClosed() name.

TODO: The remaining _implementation_ symmetry among different
purpose/scope methods (clientClosed and handleServerClosure) is a red
flag.