Cancel the close handler when retrying a closing to-server connection
AFAICT, when we decide to retry establishing a tunnel, the to-server
connection might already be in the fde::closing() state (e.g. because of
a shutdown or some other "external" event): The closing sequence has
started, our close handler has been scheduled but has not been called
yet. If we leave the closing handler "as is", and it was not a shutdown,
the closing handler will be called while we are already working on
another to-server connection -- Comm does not know that we do not care
about the old connection anymore.

It would be rather difficult for our close handler to detect and ignore
such "stale" callbacks. Thus, I see not better option than canceling
the callback instead.

This change also fixes a bug: The closer field was left dirty for such
"closing at close()"" time connections.

Since normal close() calls trigger closing callbacks, we should mark
this method specially. After doing so, I discovered that it is only
called from one place! This discovery allowed me to move the (rather
natural for a closing method) "there is no pending writer" assertion
inside the method. Without that move, we would have to cancel the
writer callback to keep all fd-related state internally consistent.