Do not send error responses after sending 200 (Connection Established)
retryOrBail() is used in more contexts than the original error sending
code its "bail" path is reusing. We not must avoid sending two HTTP
responses (one 200 (Connection Established) and one an error page).

I rejected the idea of modifying clientExpectsConnectResponse() to
become "expects NOW" rather than the current "expects as the first
(encrypted) byte" primarily because the method is used in 200-sending
context where the connection _must_ be pristine (i.e. it would be wrong
to skip sending 200 (Connection Established) response if it is dirty as
if it were a normal/possible outcome for that code context).