Fixed memory size calculation for shared memory pages on 64-bit systems
Due to code duplication, the wrong type (uint32 instead of size_t) was
used to calculate the PagePool::theLevels size. theLevels memory
(positioned by different and correct code) did not overlap with the raw
pages buffer, but the raw pages buffer could, in some cases, be 32 bits
short -- the last 4 bytes of the last page were out of allocated memory.

In practice, shared memory allocations are page-aligned, and the
difference in 4 bytes was probably compensated by the extra allocated
bytes in most (or perhaps even all) cases.