Remove custom exception handlers from Tunneler and PeerConnector
These were added/enabled in branch commits 610406b and 0e641e4. The
AsyncJob::callException() API was meant for customizing job exception
handling. Most likely, it is the result of over-engineering (and should
be removed), but what is more important in this branch project is that
we do _not_ want to handle job exceptions in a special way because the
existing swanSong() code already handles exceptions correctly.

The rationale for customizing exception handling in earlier branch
commits was to avoid BUG messages and
ERR_GATEWAY_FAILURE/Http::scInternalServerError transaction outcomes
when an "unexpected" connection closure leads to a throwing Must().
Since the previous branch commit, the known connection closure cases no
longer lead to Must() failures, and all other Must() failures should
(continue to) lead to BUG messages and
ERR_GATEWAY_FAILURE/Http::scInternalServerError outcomes (because they
are likely to be actual bugs or use cases we should at least triage).