Remove an non valid XXX comment
The comment says that if the connection is not closed after a throw
inside advanceDestination then this is will not result in
to a "BUG #3329" warning in ~Connection.

The tests shows that this is not correct, the mesage will apear after
the last owner of the connection object, the AccessLogEntry::hier::tcpServer,
be destroyed.