fixup: Undo 09ae483. Uniformity is dangerous for this code.
The `v` string must be zero-terminated because it is fed to
Cp1251ToUtf8() and Latin1ToUtf8() that expect a c-string. Thus, we
should use the termedBuf() call to initialize `v`. The other nearby
rawBuf() calls are OK because they are used in contexts that do not rely
on zero-termination.

Both String::termedBuf() and rawBuf() return the same value, so there
was no bug here, but it is best to preserve this difference because it
may become important for those who convert this code to SBuf or another
class where raw access does not return a terminated string.