Retry request if an 304 update results in too large header
When Squid updates HTTP response headers after receiving
a 304 (Not Modified) response, there are no limits on how big the
updated/cumulative header may become (and how many header fields
it may gain) across multiple updates. This accumulation leads to
and/or exposes a number of internal problems, ranging from stalled
worker processes (while Squid is slowly growing a buffer to
accommodate a huge header) to parsing errors (when storing
and/or loading cache entries).

This change fixes Squid to honor 'reply_header_max_size' on 304
updates accumulation. Now the request will be retried if the
update is going to overflow that maximum value.