HttpHeader packed size should honor configured maximums
These maximums are 'request_header_max_size' for requests and
'reply_header_max_size' for replies.

The previous approach in c364ab1 was insufficient because it fixed
only some cases of header growing (reply header updates) and because
it allowed HttpHeader object to be temporarily in an invalid state
(i.e., when it grew too large).