Remove access to clientConn after closing it
There should be no reason to dereference clientConn after bail(), but if
some (future) code reaches clientConn, it will not be tricked into
thinking that we still have access to the client Connection. For
example, current PeekingPeerConnector::initialize() uses clientConn->fd
if clientConn is set, without checking whether clientConn is open. The
latter may be a bug on its own, but it is best to avoid the ambiguous
and dangerous "exists but closing or closed" connection state.