fixup: Clarify that exclusive ownership of a Connection does not protect
... from connection closures.

The old branch comment said that we are the only ones who can close the
Comm::Connection object, which some are likely to misread as if we are
the only ones who can close the connection itself.

Moreover, the Must() before the closing() check was paranoid: We have
checked that condition in the constructor and it could not have changed
while we were waiting for start() (because we have exclusive Connection
control). I think it is safe to replace that paranoid Must() with an
assert().