Handle the case the PeekingPeerConnector::clientConn is closed
Commit 25f70a4c ("Assert that clientConn->fd usage is safe") prevent
the worst case scenario of a segfault but does not fixes squid
abort if the clientConn is closed. Unfortunately we can enter this
state after the clientConn is closed but the related ConnStateData
object is not destructed yet (its comm-close handler is not called yet).

This patch replaces the assertion with an "if" check which aborts the
procedure. Also adds a similar "if" check in noteNegotiationDone method
to protect the switchToTunnel call from similar segfaults.

A full fix may completely removes the PeekingPeerConnector::clientConn
member and use the request->clientConnectionManager->clientConnection
instead to access client connection.