Remove pointer from the input of Digest nonce hashes
This is a follow-up to #491, which hashed what was previously revealed
as plaintext. Removing the pointer from the input to the hash removes
the possibility that someone could recover a pointer by reversing a
hash. Having the pointer as input was not adding anything: Squid
remembers all outstanding nonces, so it really only requires uniqueness,
which is already guaranteed by the authenticateDigestNonceFindNonce
loop.