Remove the Ssl::ErrorDetail::request member
... to avoid cyclic depedencies after converting the Ssl::ErrorDetail
class to an ErrorDetail kid class.

The methods Ssl::ErrorDetails::toString and Ssl::ErrorDetails::buildDetail
merged to the new member Ssl::ErrorDetail::detailString. The HttpRequest
passed as argument to this method.
The new method does not caches the generated string anywhere, it is a
feature which is not actually needed, and the member errDetailStr is
removed.