The ERR_DETAIL_BUFFER error detail
Use this new error detail code to mark cases where we failed
to tunnel on error because we did not buffer any data.

Calling the tunnelOnError when we are not buffering client data
looks like a bug, however the related code is a little complex
we may have unhandled cases.