Fix SEGSEGV while composing certificate validator helper request
This is may caused if squid compiled with one of the latest openSSL
releases which report X509 errors which is not imported yet in squid.
This patch fixes the segmentation fault, by sending a single integer
error code to certificate validator if the error is unknown and also
syncs Squid certificate verification errors with latest OpenSSL
X509 errors.

This is a Measurement Factory project