fixup: Clarified why folks need to stage workspace changes.
The script cannot, technically, create "conflicts" (at least not in a
git sense of that word). The script might wipe out or corrupt unstaged
changes, and we want to protect users from that.

A very useful side effect of our protection is that any script changes
are very easy to _review_ using `git diff --word-diff`. Any developer
ought to review codespell changes. The script itself is probably not the
right place to document/discuss all that.