Handle TODO: "Explain why we should disable v1.3 support for, say, TLS v1.4"
Here we want to disable TLSv1.3 support in openSSL if the client sends
a TLSv1.2 or earlier message to the server to avoid "inappropriate fallback"
error messages when the openSSL will process later the server response.
Fix the code to check for "TLSv1.2 or earlier" instead of checking for
"not TLSv1.3" to avoid problems when the TLSv1.4 released and implemented
in OpenSSL future releases.