FwdState: keep correct prime/spare address order when retrying
Before this fix, FwdState::fail() incorrectly assumed that 'serverConn'
is always 'prime', however it may be 'spare', which is important when
re-inserting the address into the 'destinations' (ResolvedPeers) array.
We need to place 'prime' always before 'spare' for the given peer.
Ideally, we need an ability to 'return' a taken address back into
'destinations' at the same position.

The difficulty is that an address was classified as prime/spare only
within HappyConnOpener, so we could not distinguish it (and hence,
re-insert into destinations) outside the job scope. To overcome
this, we have to re-work ResolvedPeers implementation. Now the address
container does not remove a candidate address from the list but just
marks it as 'unavailable'. The client code may easily return the taken
back into the list, preserving the correct address order.