Fix Tls1p2orEarlier and Tls1p3orLater to throw on non-TLS/SSL protocol
The Tls1p2orEarlier and Tls1p3orLater now uses the TlsVersionEarlierThan
function and throws if no TLS/SSL protocol is given.
The responsibility for valid protocol is not on callers side: A check
for valid protocol is added to Ssl::ServerBio::tls1p3orLater to cover
the case the server hello message parsing is not complete because of
an error.

Also this patch adds the function Security::IsTlsProtocol to check
if the given protocol is a TLS/SSL protocol.