Improve ToNanoSeconds overflow check
Instead of multiplying and checking, which might overflow the
representation of the time unit, perform integer division
on the integer values and then compare against the
non-moltiplied supplied value.
This could mean losing a bit of space at the very top end
of std::chrono::nanoseconds::max() of up to one unit of duration
due to integer division rounding but it should be an
unlikely enough case.