fixup: Adjusted Ssl::ServerBio to reduce unknown TLS version problems
PeekingPeerConnector::noteNegotiationError(), the single tls1p3orLater()
caller, does not really care about the TLS version. It is interested in
extracting the server certificate. I adjusted the API accordingly. This
does not fully eliminate the confusion related to unknown versions, but
it makes the API simpler (fewer "or"s) and matches how existing
ServerBio methods like resumingSession() work.

Another good option would be for ServerBio to return the raw version and
let the caller figure out what to do with unknown versions.