Simplified by clearly defining currentPeerIndex and narrowing its scope
The exact meaning of currentPeerIndex was difficult for me to grasp. The
description did not tell me what the field meant. The method that was
setting it (i.e. cacheFront) did not appear to have a concept of "peer"
while the currentPeerIndex name and its users seem to make a big deal
out of that association. Given these complications, it was very
difficult for me to convince myself that the code is doing the right
thing, especially that we are not missing some currentPeerIndex updates.

I also feared that currentPeerIndex overstepping this branch scope and
doing more than just skipping unavailable candidates at the beginning of
paths_ because cacheFront(), its calculator, is called in a context that
is not directly related to availability changes or, more precisely, that
is not the only one indirectly related to availability changes.

I replaced currentPeerIndex with candidatesToSkip that has a clear,
well-defined (and now documented) semantics and that is updated when it
has to be updated based on its semantics.