Explained why we cannot easily optimize ResolvedPeers::retryPath() search
A slower retryPath() after an error response does not worry me as much
because those errors should not happen often, but I am a bit worried
that retryPath() is called for canceled spare/prime attempts which could
be happening quite often if the prime family is slow enough to trigger a
spare attempt.

Most likely, we should return an index if we expect users to essentially
undo an earlier extract*() call. That solution avoids any searching and
is also safer than a raw pointer search (from code quality p.o.v.).