Use SBuf::chop istead of SBuf::consume to avoid unneeded buffer reallocate
The SBuf::consume method creates and returns also an SBuf with the
consumed part of SBuf. This is increases temporary the lock-count
of the attached MemBlob and does not allow to clear it if all of
the data are consumed.

This is how the consume method operates in HttpStateData::inBuf case:
 - ReserveSpace of SQUID_TCP_SO_RCVBUF size for inBuf.
 - Read ReadSize<SQUID_TCP_SO_RCVBUF size of bytes.
 - Store the data to StoreEntry and consumes all the bytes from
   inBuf. The inBuf MemBlob does not cleared and now has
   (SQUID_TCP_SO_RCVBUF-ReadSize) space available
 - At the nexst step HttpStateData requires SQUID_TCP_SO_RCVBUF
   available size and the inBuf releases the current MemBlob and
   allocate a new one.
For every read HttpStateData allocates a new  inBuf buffer without
a strong reason for this.

The chop method just removes the unneeded data from the beggining
and always if no data exist in MemBlob and no other MemBlob users
exists reset it to be used again. The chop method should be used
instead of consume method if we want to just drop the consumed
data.