Improve SBuf::cow to avoid new buffer allocation in more cases
Currently the SBuf::cow reallocates a  new MemBlob even if the
current SBuf is the only owner of MemBlob and there is enough
available space. This patch increases the cases where the cow
decides to reuse the existin MemBlob.

TODO: The SBuf::cow can be more improved if we use memmove to
copy the stored bytes at the beggining of existing MemBlob buffer to
allow use all of the available space and avoid allocation of a new
MemBlob.