Avoid input buffer re-allocations while processes chunked body responses
Most of the reallocations in this case caused because if SBuf limitations
and the way we are using SBuf buffers.
This patch tries to:
  - Make the HttpStateData::httpChunkDecoder (TeChunkParser) the only owner
    of input buffer content
  - Use SBuf::chop instead of SBuf::consume inside TeChunkedParser. The chop
    operations avoids reallocation if we are the only owner.
  - Be sure that any other reference to TeChunkedParser::buf_ content
    is removed to allow SBuf::chop operation avoid reallocation if possible.

XXX: This is a patch mostly to show the SBuf limitations and the
problematic way we are using it. This patch or similar maybe can
be accepted as optimization but probably a more complete patch
should developed to rewrite parts of Http parsers.