fixup: Optimized ResolvedPeers::size()
The method is not used on the performance sensitive paths, but I was
worried about its linear search in debugging runs. It was easy to
optimize and the stored information may assist in debugging sessions.

New fast size() allowed me to undo changes to the ResolvedPeers printing
operator and to remove the const version of start().