Initial support for --keep-going
To simplify problem triage and to follow common approach used by other
development tools that apply multiple actions to sources (e.g., make and
gcc), we should stop previous error(s) may affect the next step. We
apply this principle to the being-added plugin support (for now).

Folks that want to apply as many source maintenance modifications as
possible, regardless of individual modification failures, should run
scripts/source-maintenance.sh --keep-going

Since plugins are not supposed to have state, we could continue applying
plugins to source files that were not affected by errors -- doing so
will not significantly complicate triage because the problematic files
are left intact after the first failure. This would be conceptually
similar to a compiler reporting more than one error for a given source
file before refusing to process that file further.

However, I think it is best to wait with that enhancement until we have
a better understanding (including a variety of real examples) of
plugins. For example, it is possible that the plugin interface will be
adjusted to supply the file name to the plugin. In that case, a plugin
could consult other source files (e.g., look at the .h header for a .cc
source file), essentially gaining access to state.