Optimized error detail generation and reduced diff noise
We do not need to create the same ErrorDetail object many times. We can
use a constant object instead. And naming that constant object using old
ERR_DETAIL_* enum constants allows us to remove a lot of non-changes
from the diff -- most of the high-level changes in this commit are
undoing earlier branch changes.