Sketched proper SSL error handling in tlsAttemptHandshake()
This commit adds a few lower-level XXXs that still need to be addressed,
but I wanted to capture the essence of the change: The high level code
should capture SSL-related errors from all levels and then let
Ssl::ErrorDetail report them as needed. The previous approach was trying
to decide which low-level errors to capture, which led to many code
inconsistencies/questions and probably led to some details being lost.