Decided how to handle multiple error details for the same ErrorState
The first-generated detail should be used. This was a relatively easy,
natural decision after I realized (and documented) that we cannot
combine multiple details in one ErrorDetail -- parts of those details
overlap.

It was more difficult to decide how to enforce the right priority. I
settled on a simple (but imperfect) solution -- tracking which
ErrorDetail object was generated first. This only works if detail
objects are never created before the error discovery.

I rejected the idea of always setting the previously-nil
ssl_ex_index_ssl_error_detail when generating any Ssl::ErrorDetail
because I was not sure that the TLS connection can be provided to all
ErrorDetail contexts, especially built-in handshake parsing contexts. I
was also worried that it would lead to more complicated final code and
an increase in error handling overheads. On the other hand, the rejected
approach, if successful, would have offered firmer "first details wins"
guarantees by providing a single synchronization medium.

Also fixed an earlier branch bug: One of the Ssl::ErrorDetail ctors
forgot to set error_no.