Undo most ConnStateData::requestTimeout() branch changes
Our goal is to detail errors, not the state of a transaction during an
error. The latter may be very valuable, but is out of scope and should
probably be done differently: Instead of trying to restore the current
state based on various artifacts like SSL_is_init_finished() and
fd_table[].ssl, the transaction should remember its last/current step(s)
that we should log (using a new %code?).

I hope none of the known triage cases relied on the ErrorDetail
information that this commit removes. If they did, we may need to
consider opening a separate project to detail transaction state.