Merged err_type and ErrorDetail into the new Error class
All classes storing err_type either also store ErrorDetail or should
store ErrorDetail (but have not been introduced to that type yet). In
this branch, there was an increasing number of classes that needed to
store both pieces of the error-related information, and error
storage/update code was being duplicated across those classes.

I placed the new class (and related code) into the new src/error/
directory primarily because it makes it easy to create error/forward.h
but also because we should probably split SysErrorDetail and
ExceptionErrorDetail into their own files (and we should not pollute
src/ with those; we were already polluting with the new ErrorDetail.*).

I did not move errorpage into src/error/ to reduce distracting changes
(this branch does not modify that file much) and because I was worried
that placing such a high-level class into that (currently low-level)
library will cause linking headaches, especially for unit tests.

I did not merge ErrorState::type and ErrorState::detail to avoiding
"distracting changes" and because properly merging them into
ErrorState::error would require to get rid of ErrorState::xerrno as
well, introducing even more out-of-scope changes. Left a TODO for that.