Remove extra Entry allocation
Entry can be safely stored in the Map sub-container
and have unchanging references/pointer stored in Queue
for the duration of the Entry data being stored in this
ClpMap container.

This reduces three allocations (Queue, Map, Entry) to
two (Map+Entry, Queue) and removes the need for
managing volatile Queue::iterator's.

Also, optimize Map storage management with emplace()
and move semantics on Entry.