Fix Linux build with glibc v2.30 which deprecated sys/sysctl.h
    error: #warning "The <sys/sysctl.h> header is deprecated and
    will be removed." [-Werror=cpp]

Squid only uses sysctl() on FreeBSD and friends so let's not #include
sys/sysctl.h in other environments. The existing compat/shm.cc already
uses a similar condition for including this header.

Glibc recommends using /proc instead of the sysctl.h they deprecated,
but that is probably not an option for FreeBSD anyway.