RFC 3986 compliant relative-reference logic
Update the urlIsRelative() test from RFC2396 to RFC3986. The new
criteria are more precise about what is a reference and what is
not. Also about query-string and fragment relevance.

Update AnyP::Uri::addRelativePath to merge '/path' references
as replacement only for the last segment, not the entire path.
The current only caller no longer needs this custom hack.

Normalization of . and .. URL are still not handled properly.
The resulting path_ string will not be fully normalized after
merge.