Encapsulated application of plugin(s) to a given source file
I did not name the new function applyPlugins() because a single-letter
difference with the existing applyPlugin() seemed too dangerous,
especially since the two functions have a different parameter order.

I did not change applyPlugin() parameter order because the call would
look a little odd/out-of-order and because it is possible that
applyPlugin() will be eventually used to apply a single plugin to
multiple sources.

TBH, I did both first, but did not like the results.