fixup: Rephrased comments to deemphasize GREASE value concerns
A GREASE value is supposed to be treated just like any other unsupported
value so mentioning GREASE explicitly every time we have to deal with a
possibly-unsupported value does not make much sense. A GREASE value is
just one common (in some contexts) example of an "unsupported" value.

Also emphasize (a little) that "supported" is relative to the
agent/context. Many of the TLS values that Squid does not know about are
supported by the TLS client or server that Squid is peeking at.