fixup: Emphasized that ParseOptionalProtocolVersion() encodes unsupported
... versions. In hope to address Amos review concern.

We should not move the "ignore" text to the parser because the parser
does not ignore anything. It only hides the details (which was
emphasized in the previous branch commit). The "In either case, we would
continue ignoring all the unsupported values here" phrase cannot be
moved to the parser because it is specific to this ignoring code. Other
future callers may not ignore unsupported values but do something
special with them (e.g. store them for logging).