Simplified Security::ErrorDetail certificate management
This simplification not only improves code clarity but allows a
TLS-library-free build to succeed. The added cost of an extra
certificate lock in the former absorbPeerCertificate() context is a fair
price to pay for these significant improvements IMO.

Also do not implicitly rely on the private ErrorDetail constructor to
not set errReason and certificates. The certificate validation
constructor makes the final decision regarding those fields.

This commit requires a followup fix to support nullptr
paramters for CertPointer arguments. See the next commit.