fixup: Allow using nullptr for a LockingPointer formal parameter
This commit is required for the previous commit to compile. I am
segregating these changes because most of them look out of scope (even
though virtually all of them are either required or become necessary
after the required changes). It is possible to avoid all these
(positive) changes if the callers in the previous commit use
CertPointer(nullptr) instead of nullptr, but I kept these changes
because they are simple and overall positive.

CertValidationResponse response code needs more out-of-scope work.