Documented a clientReplyContext::handleIMSReply() problem
When collapsed revalidation is enabled, handleIMSReply() might get a
reply from a very old transaction, but the method continues to assume
that it is dealing with a response to the current revalidation request.
I am not sure how to detect irrelevant revalidation responses because
many (relevant) collapsed revalidation responses are received in
reaction to requests other than the current one. Such response sharing
is the whole point of collapsing requests!

clientReplyContext::cacheHit() faces a similar problem with collapsed
forwarding, but that method has never assumed that the response is fresh
-- it has always validated whether the found "hit" response is still
usable. The same cache freshness checks are unlikely to be usable "as
is" for revalidation responses. For example, Squid assumes that 304
responses are not cached. We probably need to develop a set of similar
heuristics dedicated to revalidation responses and then see what
existing code can be reused to check those heuristics.