fixup: Rename checkPeeked() to remove the implication that the checks
... are applied to the peek() result without any further alterations.
In reality, some of the peek() results also go through allowSharing()
before they are checked. The order or, more precisely, the exact
composition of those two steps (allowSharing and checkFoundCandidate)
may be revised later -- we kind of want to do everything "at once".

Also moved checkTransients() into checkFoundCandidate() because the
latter method is meant to accumulate all the post-peek() checks outside
of allowSharing().