Wrote proper per-context memory usage reporting code
The reported values will continue to be bogus, of course, until the
underlying problem -- Squid's inability to calculate the individual
context memory usage -- is addressed. That old problem is already
documented in MemoryUsedByContext(). That problem affects a lot of
things, but the code reporting context cache statistics has nothing to
do with that problem. Now it is no longer affected by it.