fixup: Improved Entry::expires description
We (usually) do not remove the entries when they expire -- ClpMap uses
opportunistic garbage collection, removing expired entries when they are
referenced or become capacity victims.

Despite the name, the stored time is not the time when the entry
actually expires (or "to be removed" if we were removing expired entries
ASAP). The entry is deemed fresh until the next second. TODO: Find a
better data member name.

Also added an (out of scope) TODO.