Report SMP store queues state (mgr:store_queues)
The state of two Store queues is reported:

Transients notification queue (a.k.a. Collapsed Forwarding queue) and
SMP I/O request queue (used for communication with rock diskers).
Each worker and disker kid reports its view of that kid's incoming and
outgoing queues state, including a small sample (up to 7 items) of
queued messages. The reports are YAML-compliant.

With the exception of a single field labeled "other", each queue report
is self-consistent despite accessing data shared among kids -- the
reported combination of values did exist at the snapshot collection
time.

The special "other" field represents a message counter maintained by the
other side of the queue. In most cases, that field will be close to its
correct value, but, due to modifications by the other process of a
non-atomic value, it may be virtually anything. We think it is better to
report (without officially naming) this field because it may be useful
in triage regardless of these caveats. Making the counter atomic just
for these occasional reports is not worth the performance overheads.