Undo finishedWithClient() changes except the argument-related fixes
Passing stale size to (and checking passed size in) the official
connFinishedWithConn() method was clearly a bug. I preserved the
corresponding fix and polished the method profile again. Technically,
this fix is out of this project scope, but it is small, was discovered
in the middle of this project touching related code, so it is OK to
include it IMHO.

I removed the other changes because they were done in the wrong method.
The connFinishedWithConn() method is about deciding whether to close the
half-closed connection. It is not about detailing any errors.

Error detailing should happen where problems are encountered. In our
context, two kinds of problems are encountered in doClientRead():

* Read error: Handled via pipeline.terminateAll(). Needs a bit more work
  to actually store the error detail -- the _ABORTED suffix is added,
  but the errno is lost. This has nothing to do with half-closures.

* Fatal EOF: Handled via connection closure, which triggers swanSong()
  which then calls pipeline.terminateAll(). Needs more work to detail
  TLS handshake parsing state -- only the _ABORTED suffix is added now.

I will add the missing bits in the next commit.