Record unexpected client connection closure problems when they happen
... so that we do not have to guess what happened in some general
cleanup method that does not have access to the error context.

This commit adds the primary missing bits identified in the previous
commit. Propagating error details to all the pending pipelined
transactions is not trivial, and this commit does not solve all the
problems.

Calling pipeline.terminateAll() on EOF creates an extra access.log entry
but this is actually not a new problem: The old code has been calling
terminateAll() twice on read errors already -- the second call always
happens in swanSong(). I hope to solve that and related "terminate once"
problems in the next commits.