Consume unparsed input attributed to a transaction; checkLogging() once
Due to fake CONNECT requests (for client TLS handshake parsing) and
from-client read errors (that occur while we are waiting for more buffer
space to read HTTP request body into), and possibly other complications,
we may have a transaction that did not consume every input byte
available to it. That transaction is still responsible for reporting
those unparsed bytes (e.g., by logging the number of bytes read on a
connection and the number of parsed bytes). It is pointless to create
another transaction -- it would not be in a better position to log them!

The above realization allows terminateAll() to consume such attributed
unparsed input, silencing subsequent checkLogging() call(s).

Calling checkLogging() many times needlessly complicates things. The
method is meant as a last-minute check to ensure that any activity worth
logging has been logged (or is forcefully logged by the method) before
ConnStateData object is gone. Thus, we should finish all transactions
(peacefully or otherwise) and then check, once, for any leftovers. If we
are checking after all transactions are gone, we do not need to guess
what any pending transactions are still doing (e.g. whether the unparsed
bytes belong to them).

Also forward bareError to the first regular transaction added to the
pipeline or the fake last-resort transaction in checkLogging() instead
of forgetting about the error.