fixup: Do not make integer parameters const in declarations
Making formal integer parameters `const` in function declarations does
not help the caller and, hence, some linting tools complain about the
"extra" `const` in the declaration. Using the same `const` in the method
definition is a good idea. Personally, I do not like seeing a method
declaration that differs from the method definition but, AFAICT, that
approach is overall better than alternatives, so I use it.