Count cache entry key memory consumption once
When ClpMap is used for sslcrtvalidator_program, the key is the whole
validator request, which includes the PEM certificate. The keys I
observed in basic tests were 5KB and 10KB long (while the Value cost is
still incorrectly estimated as 72 bytes). Doubling those large values
when both SBufs necessarily point to the same MemBlob object felt wrong,
and I did not want to spread the problem around by using the extra
length to compensate for the broken response/Value size estimation.