Renamed ttl=none to ttl=infinity
"Time to live is none" means the opposite of what this setting does. The
word "none" works well for various maximums and limits, but TTL is
spelled out and, hence, interpreted by most humans as a duration when
something is valid (rather than a usability time limit). Underneath
(i.e. when interpreted by computers), both are the same, of course.